+++ R/ichimoku-package.R
@@ -23,11 +23,12 @@
 #'     The most recent version of the package may be found at
 #'     \url{https://github.com/shikokuchuo/ichimoku/}
 #'
-#' @importFrom RcppRoll roll_maxr roll_minr
+#' @useDynLib ichimoku, .registration = TRUE
 #' @importFrom ggplot2 ggplot aes geom_ribbon geom_line geom_linerange geom_rect
 #'     scale_color_manual scale_fill_manual guides scale_x_datetime
 #'     scale_x_continuous scale_y_continuous labs theme_light theme element_rect
 #'     element_line element_text
+#' @importFrom Rcpp sourceCpp
 #' @importFrom rlang .data
 #'
 #' @docType package